Split the extractor+deskewer into fs-oblivious and fs-aware pieces...
...for the wasm build.

+ dummy function in cimbar_recv_js
+ add extractor to the link closure

I've soured on the "Simple*" naming scheme though, I think I'm going to
rename these.
