+++ CMakeLists.txt
@@ -65,6 +65,7 @@ set (PROJECTS
 	src/lib/cimbar_js
 	src/lib/compression
 	src/lib/encoder
+	src/lib/extractor
 	src/lib/fountain
 	src/lib/gui
 	src/lib/image_hash
@@ -83,7 +84,6 @@ set (PROJECTS
 if(NOT DEFINED USE_WASM)
 set(PROJECTS
 	${PROJECTS}
-	src/lib/extractor
 
 	src/exe/cimbar
 	src/exe/cimbar_extract
