+++ src/lib/extractor/SimpleDeskewer.h
@@ -1,0 +1,38 @@
+/* This code is subject to the terms of the Mozilla Public License, v.2.0. http://mozilla.org/MPL/2.0/. */
+#pragma once
+
+#include "Corners.h"
+
+#include "util/vec_xy.h"
+#include <opencv2/opencv.hpp>
+
+#include <vector>
+
+class SimpleDeskewer
+{
+public:
+	SimpleDeskewer(cimbar::vec_xy image_size={}, unsigned anchor_size=0);
+
+	template <typename MAT>
+	MAT deskew(const MAT& img, const Corners& corners);
+
+protected:
+	cimbar::vec_xy _imageSize;
+	int _anchorSize;
+};
+
+template <typename MAT>
+inline MAT SimpleDeskewer::deskew(const MAT& img, const Corners& corners)
+{
+	std::vector<cv::Point2f> outputPoints;
+	outputPoints.push_back(cv::Point2f(_anchorSize, _anchorSize));
+	outputPoints.push_back(cv::Point2f(_imageSize.width() - _anchorSize, _anchorSize));
+	outputPoints.push_back(cv::Point2f(_anchorSize, _imageSize.height() - _anchorSize));
+	outputPoints.push_back(cv::Point2f(_imageSize.width() - _anchorSize, _imageSize.height() - _anchorSize));
+
+	MAT output(_imageSize.height(), _imageSize.width(), img.type());
+	cv::Mat transform = cv::getPerspectiveTransform(corners.all(), outputPoints);
+
+	cv::warpPerspective(img, output, transform, output.size(), cv::INTER_LINEAR);
+	return output;
+}
