+++ src/lib/extractor/SimpleExtractor.cpp
@@ -1,13 +1,13 @@
 /* This code is subject to the terms of the Mozilla Public License, v.2.0. http://mozilla.org/MPL/2.0/. */
-#include "Extractor.h"
+#include "SimpleExtractor.h"
 
-#include "Deskewer.h"
+#include "SimpleDeskewer.h"
 #include "Scanner.h"
 #include "cimb_translator/Config.h"
 #include <vector>
 using std::string;
 
-Extractor::Extractor(cimbar::vec_xy image_size, unsigned anchor_size)
+SimpleExtractor::SimpleExtractor(cimbar::vec_xy image_size, unsigned anchor_size)
 	: _imageSize({
 		image_size.width()? image_size.width() : cimbar::Config::image_size_x(),
 		image_size.height()? image_size.height() : cimbar::Config::image_size_y()})
@@ -15,7 +15,7 @@ Extractor::Extractor(cimbar::vec_xy image_size, unsigned anchor_size)
 {
 }
 
-int Extractor::extract(const cv::Mat& img, cv::Mat& out)
+int SimpleExtractor::extract(const cv::Mat& img, cv::Mat& out)
 {
 	Scanner scanner(img);
 	std::vector<Anchor> points = scanner.scan();
@@ -23,7 +23,7 @@ int Extractor::extract(const cv::Mat& img, cv::Mat& out)
 		return FAILURE;
 
 	Corners corners(points);
-	Deskewer de(_imageSize, _anchorSize);
+	SimpleDeskewer de(_imageSize, _anchorSize);
 	out = de.deskew(img, corners);
 
 	if ( !corners.is_granular_scale(_imageSize) )
@@ -31,7 +31,7 @@ int Extractor::extract(const cv::Mat& img, cv::Mat& out)
 	return SUCCESS;
 }
 
-int Extractor::extract(const cv::UMat& img, cv::UMat& out)
+int SimpleExtractor::extract(const cv::UMat& img, cv::UMat& out)
 {
 	Scanner scanner(img);
 	std::vector<Anchor> points = scanner.scan();
@@ -39,29 +39,10 @@ int Extractor::extract(const cv::UMat& img, cv::UMat& out)
 		return FAILURE;
 
 	Corners corners(points);
-	Deskewer de(_imageSize, _anchorSize);
+	SimpleDeskewer de(_imageSize, _anchorSize);
 	out = de.deskew(img, corners);
 
 	if ( !corners.is_granular_scale(_imageSize) )
 		return NEEDS_SHARPEN;
 	return SUCCESS;
-}
-
-int Extractor::extract(string read_path, cv::Mat& out)
-{
-	cv::Mat img = cv::imread(read_path);
-	cv::cvtColor(img, img, cv::COLOR_BGR2RGB);
-	return extract(img, out);
-}
-
-int Extractor::extract(string read_path, string write_path)
-{
-	cv::UMat img = cv::imread(read_path).getUMat(cv::ACCESS_FAST); // cv::USAGE_ALLOCATE_SHARED_MEMORY would be nice...;
-	cv::cvtColor(img, img, cv::COLOR_BGR2RGB);
-
-	int res = extract(img, img);
-
-	cv::cvtColor(img, img, cv::COLOR_RGB2BGR);
-	cv::imwrite(write_path, img);
-	return res;
 }
